requirements/base.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
openai
tzlocal
